Shader Utilities
A Resource loader class has been added to read shader files obtaining a
string which represents the source code of the shader.

What's more a shader class is now able to perform all the needed shader
operations for the program (shader addition/compilation/linking and
binding them). However no shaders are used in the project so far.