Simple Transformations
A transformation class has been added to give support for simple
transformations for now (translation) by using the previously implement
uniform variable passing to shaders.

Some methods have been added in auxiliary classes to ease the working of
the transformation one.

A small example of armonic translation is included in the main program.